Fix compilation on ARM with C++11: narrowing conversion
Cannot have double-to-float conversions inside { } in C++11.

+++ src/gui/painting/qpainterpath_p.h
@@ -268,7 +268,7 @@ inline void QPainterPathData::maybeMoveTo()
     }
 }
 
-#define KAPPA 0.5522847498
+#define KAPPA qreal(0.5522847498)
 
 
 QT_END_NAMESPACE
